tags crud

+++ src/main/java/com/meowmere/main/repositories/character/TagRepository.java
@@ -1,6 +1,7 @@
 package com.meowmere.main.repositories.character;
 
 import com.meowmere.main.entities.characters.Image;
+import com.meowmere.main.entities.characters.Relation;
 import com.meowmere.main.entities.characters.Tag;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -10,5 +11,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TagRepository extends JpaRepository<Tag, Long> {
+public interface TagRepository extends JpaRepository<Tag, Integer> {
+
 }
